enhance API & update API routes
